Task 7- stored procedure

+++ 0x0C-MySQL_Advanced/7-average_score.sql
@@ -3,7 +3,7 @@
 DELIMITER |
 CREATE PROCEDURE ComputeAverageScoreForUser(user_id INT)
 BEGIN
-DECLARE prom INT;
+DECLARE prom FLOAT;
 SET prom = (select AVG(score) from corrections AS newTable WHERE newTable.user_id=user_id;);
 UPDATE users SET average_score=prom WHERE id=user_id ;
